tests/KGP.Tests/Frames/LongExposureInfraredFrameDataTests.cs
using System;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using KGP.Frames;

namespace KGP.Tests
{
    [TestClass]
    public class LongExposureInfraredFrameDataTests
    {
        [TestMethod]
        public void TestConstrutor()
        {
            LongExposureInfraredFrameData data = new LongExposureInfraredFrameData();
            bool pass = data.DataPointer != IntPtr.Zero;
            data.Dispose();
            Assert.AreEqual(true, pass);
        }

        [TestMethod]
        [ExpectedException(typeof(ObjectDisposedException))]
        public void TestDispose()
        {
            LongExposureInfraredFrameData data = new LongExposureInfraredFrameData();
            data.Dispose();
            Assert.AreEqual(data.DataPointer, IntPtr.Zero);
        }

        [TestMethod]
        [ExpectedException(typeof(ObjectDisposedException))]
        public void TestDisposeAccess()
        {
            LongExposureInfraredFrameData data = new LongExposureInfraredFrameData();
            data.Dispose();

            //Should throw exception
            var pointer = data.DataPointer;
        }

        [TestMethod]
        [ExpectedException(typeof(ObjectDisposedException))]
        public void TestMultipleDispose()
        {
            LongExposureInfraredFrameData data = new LongExposureInfraredFrameData();
            data.Dispose();
            //Second call to dispose should do nothing
            data.Dispose();
            
            Assert.AreEqual(data.DataPointer, IntPtr.Zero);
        }

        [TestMethod]
        public void TestSize()
        {
            LongExposureInfraredFrameData data = new LongExposureInfraredFrameData();
            int expected = 512 * 424 * 2;
            bool pass = data.SizeInBytes == expected;
            data.Dispose();
            Assert.AreEqual(pass, true);
        }

        [TestMethod]
        public void TestDisposedSize()
        {
            LongExposureInfraredFrameData data = new LongExposureInfraredFrameData();
            data.Dispose();
            Assert.AreEqual(data.SizeInBytes, 0);
        }
    }
}
